requirements.txt
# Generated from requirements.in via pip-compile.
# To regenerate: pip-compile requirements.in --output-file requirements.txt

aiohttp==3.10.10
    # via requirements.in
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2
    # via httpx
async-timeout==4.0.3
    # via aiohttp
attrs==24.2.0
    # via aiohttp
certifi==2024.8.30
    # via httpcore, httpx
cffi==1.17.1
    # via pynacl
discord.py[voice]==2.4.0
    # via requirements.in
frozenlist==1.4.1
    # via aiohttp, aiosignal
h11==0.14.0
    # via httpcore
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via requirements.in
idna==3.10
    # via anyio, httpx
multidict==6.1.0
    # via aiohttp, yarl
propcache==0.2.0
    # via aiohttp, yarl
pydantic==2.9.2
    # via pydantic-settings
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.6.1
    # via requirements.in
pynacl==1.5.0
    # via discord.py
python-dotenv==1.0.1
    # via pydantic-settings
redis==5.2.0
    # via requirements.in
sniffio==1.3.1
    # via anyio
typing-extensions==4.12.2
    # via pydantic, pydantic-core
yarl==1.17.1
    # via aiohttp
